feat: Enhance ticket management with new consumers for releasing and reserving tickets; update OpenTelemetry configuration; improve logging in PaymentService and TicketService; refactor BookingStateMachine for better error handling

+++ services/SharedLibrary/Contracts/Messages/Messages.cs
@@ -3,7 +3,7 @@
 public record HoldTickets
 {
     public Guid BookingId { get; init; }
-    public int EventId { get; init; }
+    public long EventId { get; init; }
     public int NumberOfTickets { get; init; }
     public string CustomerId { get; init; } = string.Empty;
 }
@@ -19,7 +19,7 @@ public record ProcessPayment
 public record ReserveTickets
 {
     public Guid BookingId { get; init; }
-    public int EventId { get; init; }
+    public long EventId { get; init; }
     public List<Guid> TicketIds { get; init; }
     public string CustomerId { get; init; } = string.Empty;
 }
@@ -27,7 +27,7 @@ public record ReserveTickets
 public record ReleaseTickets
 {
     public Guid BookingId { get; init; }
-    public int EventId { get; init; }
+    public long EventId { get; init; }
     public List<Guid> TicketIds { get; init; }
     public string Reason { get; init; } = string.Empty;
 }
